modified to work with Eachine VR D2
Made a lot of changes to support the Eachine VR D2 goggles
including support for the 2 digit display.

Changed the button handling to work better with only 2 buttons.

Added custom code to handle the two buttons wired to one pin.
The two pins the select which digit to display also select
which button press is being monitored on the input pin.

Cleaned up the loop function.