Eachine VRD2 channel switching with mosfets
On short CAMDVR button press, and if enabled in the settings, switch
between CAM, DVR and OSD modes internally (i.e. without changing between
AV1/AV2 at the display) using two mosfets connected to A2 and A3.
The A2 mosfets has the gate pin connected to A2, the drain pin to the
receiver output, i.e. what previously went to the display as AV1, and
the source pin connected to the Pro Mini's OSD output together with the
other mosfet's source pin.  The two source pins and the OSD output then
go out to the display as either AV1 or AV2.  At any time only one mosfet
may be enabled or both must be disabled if the OSD is active.  If OSD is
inactive the Pro Mini's D7 and D9 pins are set to inputs.